requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5

# Optional: better detection on varied lighting / cluttered scenes (pip install ultralytics)
# ultralytics>=8.0.0
